23.06.03 - 8. Commit | Process: Upon the user's selection of a fragment on the main page, the headerText is styled with white text color, a pink background, and a rounded corner radius. The sorting process for displaying products with approaching expiration dates on the main page based on the closest date has been completed. The Settings section has been updated as Expenses, where the user's expenses for the past month are listed.

+++ app/src/main/java/com/mustafagur/marketim/DetailActivity.kt
@@ -71,7 +71,7 @@ class DetailActivity : AppCompatActivity() {
         builder.setMessage(""+urunadi+" isimli ürünü silmek istediğine emin misin?")
         builder.setPositiveButton("Evet") { dialog, which ->
             val db = DatabaseHelper(this)
-            db.deleteData(urunid.toLong())
+            db.deleteData(urunid)
             db.close()
             Toast.makeText(this,"Başarıyla Ürün Silindi!",Toast.LENGTH_LONG).show()
             val go = Intent(this@DetailActivity, MainActivity::class.java)
